refactor: implement useAsync hook for simplified async handling
